merge: fold realsense_depth_map_acc into realsense_depth_map, drop DepthStreamer
